Refactor #85 : 마일스톤 전체 조회 res 형식 변경
- milestone의 갯수를 알려주는 count 속성 추가
- service try 문 안으로 return 이동
