+++ server/src/api/index.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const userRouter = require('@api/user/user-routes');
 const issueRouter = require('@api/issue/issue-routes');
+const milestoneRouter = require('@api/milestone/milestone-routes');
 const { apiAuth } = require('@middlewares/auth');
 
 const router = express.Router();
@@ -10,5 +11,8 @@ router.use('/user', userRouter);
 router.use(apiAuth);
 
 router.use('/issue', issueRouter);
+
+
+router.use('/milestone', milestoneRouter);
 
 module.exports = router;
