+++ server/src/api/milestone/milestone-routes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 
-const milestoneController = require('@api/user/milestone-controller');
+const milestoneController = require('@api/milestone/milestone-controller');
 
 const router = express.Router();
